arch/amd64: Implement mmu_native_init()
After the code has bootstrapped itself to kmain(), it's still using
the boot page directories which are really not mean to be used for
anything else than booting.

Identity map the lower part of the kernel (that was used during booting)
bot also an additional 2GB of address space for the kernel.

The virtual address space for kernel starts at 0xffffffff80100000
and the only strict upper bound for it is 0xffffffffffffffff.