mm: Add initial kernel heap implementation
Kernel heap is used to allocate the SLAB and page frame allocators
so it will be initialized twice: once during MMU initializations using
the boot memory allocator and once using the page frame allocator.

For now the heap has a very simple design where the memory (both free
and in-use) are divided into arenas which are further divided into
linked list of chunks. kmalloc()/kzalloc() traverses through these
arenas and chunks in search of a memory block large enough to satisfy
the call and if no block is found, more memory is allocated.