mm: Implement page frame allocator
The page frame allocator is implemented as buddy memory allocator
which allows the caller to allocate a block of physical memory. The
smallest unit is 4 KB and the largest unit is 65 KB.

The PFA is implemented in two steps. Firstly it claims all available
address space that Multiboot has reported, and this is called the
pre-init phase. It then allocates itself a chunk of memory and initializes
an array representing the found physical memory. This is the
initialization phase. When the initialization is ready and there is a
record of all available physical memory, the PFA executes the post-init
phase where it marks the pages it has allocted for itself as occupied.