fs: Add pipe support
Pipes are used by, e.g., the PS2 driver for handling user input.

Currently the implementation is lacking the scheduling-related parts so
it returns `-ENOSPC` instead of blocking on the pipe as right now there
is no tasking support implemented.

In the future the pipe object shall have a wait queue which it uses
to suspend the reading/writing task if the pipe is empty/full.