kernel: Fix how IRQs are handled
Some PCI devices may share the same IRQ pin which means that the
interrupt handler may have to call multiple handlers in order to
to handle the interrupt. This is why the interrupt handlers
return `IRQ_HANDLED/IRQ_UNHANDLED` to indicate whether this handler
on the IRQ pin actually received an interrupt.

Loop over all installed handlers for the the IRQ pin that issued an
interrupt until one of the handlers returns `IRQ_HANDLED`.